fix: circular dependency between nginx, headscale and tailscale
For tailscale to work the headscale server must be accessible, for the
headscale server to be accessible nginx must be active and proxying it,
but for nginx to start it needs to bind to an internal address so it can
proxy internal services which failes as tailscale isn't running yet so
these addresses don't exist.

Fix this by only allowing access from tailnet addresses. Don't really
know why I didn't do this in the first place.